Fix runtimes init (#39)
* Fix flawed assumption in runtimes.init

It assumed there would be a valid current namespace in the credential
store.  This is usually true in normal use but when nim is embedded
in a runtime container to support remote build it is not true.

* Export initRuntimes instead of init

Every use site in nimbella-cli has to rename at import time to avoid
confusion.  We shouldn't be exporting such a generic name in the first
place for such a specific function.

* 3.0.6