Runtimes with a "default build" cause builds to be remote by default (#77)
* Languages with default build now remote by default

Now the only way to suppress this behavior is by setting the localBuild
flag for the action in project.yml.  Adding a command line flag to
force local build seems unnecessary but may be considered later.

* Fix logic error and some breakages.

The fact that `hasDefaultRemote` is now called in some cases where it
wasn't before revealed some fragile assumptions that needed to be
hardened.