New package*.json files address vulnerabilities (#79)
* New package and lock files address vulnerabilities

There has meanwhile been neglect of some of the dev dependencies which
needed to be updated to support a from-scratch npm install.  There are
complex peer-dependency interactions between esbuild and jest and some
changes to jest.

* Also drop 12.x from automated workflows

The new esbuild may be incompatible with 12.x.  But, as the engine
declaration is at 14.x.x we should be dropping 12.x from the tests
anyway.