Several changes, leading to a candidate deployer:
1.  Command is renamed to 'dosls'.  We don't expect to use this command
except for testing and perhaps inside runtimes, where the name won't
matter.
2.  Minor refactoring so most of the 'main' logic will be reusable in a
reduced 'doctl' plugin.  This is how almost all users (including most
internal ones) will consume the remaining logic.
3.  Add (probably temporary) support for 'watch'.  It is likely that we
will (after all) recode the actual watching in golang and just call the
plugin when there is something to deploy.  We might add a '--watching'
flag to cover the subtle differences in output that we need in watching
vs a simple deploy.